+++ Sources/Models/Database/Chapter.swift
@@ -10,10 +10,11 @@ public final class Chapter {
     public var content: String? // Đối với truyện tranh, content chứa danh sách ảnh cách nhau bởi dấu xuống dòng (\n)
     public var isCached: Bool = false
     public var titleTrans: String?
+    public var host: String?
     
     public var book: Book?
     
-    public init(id: String, title: String, url: String, index: Int, content: String? = nil, isCached: Bool = false, titleTrans: String? = nil) {
+    public init(id: String, title: String, url: String, index: Int, content: String? = nil, isCached: Bool = false, titleTrans: String? = nil, host: String? = nil) {
         self.id = id
         self.title = title
         self.url = url
@@ -21,5 +22,6 @@ public final class Chapter {
         self.content = content
         self.isCached = isCached
         self.titleTrans = titleTrans
+        self.host = host
     }
 }